backend/cv-renderer-service/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
pydantic==2.12.5
python-docx==1.1.2
